Move API call methods to their own file
