+++ src/components/Navigation.jsx
@@ -1,5 +1,5 @@
 import React, {useState , useEffect} from 'react';
-import axios from 'axios';
+import { getCharts, saveChart } from '../helpers/chartRequests.js';
 
 var Navigation = (props) => {
 
@@ -47,53 +47,3 @@ var Navigation = (props) => {
 }
 
 export default Navigation;
-
-var getCharts = (usernameInput,updateSavedCharts) => {
-  axios.get('/charts', {params: {
-    username: usernameInput
-    }
-  })
-    .then((response) => {
-      console.log('get chart response data', response.data)
-      updateSavedCharts(response.data)
-    })
-    .catch((err) => {
-      console.log('error retrieving charts')
-    })
-}
-
-var saveChart = function (username, chartName, assets, labels, dataPoints, categories, savedCharts, updateSavedCharts) {
-
-  if (!username.length || typeof username !== 'string') {
-    alert('Invalid username');
-    return;
-  }
-
-  if (!chartName.length || typeof chartName !== 'string') {
-    alert('Invalid chart name');
-    return;
-  }
-
-  var newChartEntry = {
-    username: username,
-    chartname: chartName,
-    labels: labels,
-    assets: assets,
-    dataPoints: dataPoints,
-    categories: categories
-  }
-
-  console.log(newChartEntry);
-
-  return axios.post('/charts', { newChart: newChartEntry })
-    .then(() => {
-      console.log('saved the chart');
-      var savedChartsCopy = JSON.parse(JSON.stringify(savedCharts));
-      savedChartsCopy.push(newChartEntry);
-      console.log('savedChartsCopy',savedChartsCopy)
-      updateSavedCharts(savedChartsCopy);
-    })
-    .catch((err) => {
-      console.log('error saving the chart',err)
-    })
-}
